backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
python-docx==1.1.0
PyPDF2==3.0.1
Pillow==10.1.0
aiofiles==23.2.1
email-validator==2.1.0
python-dateutil==2.8.2
openpyxl==3.1.2
jinja2==3.1.2
